Solve Day 2 Problem 2
Just like Day 2 Problem 1 but now need to find two numbers in each row
that cleanly divide.

At first, it seemed I'd need to check every number in the row vs every
other number. However, I realized that numbers would only be divisible
by numbers smaller than themselves. This meant that if I sorted the
numbers in descending order, each number would only need to be checked
against numbers *after* it in the list.

The pattern of comparing the head against the rest of the list, then
taking the head of the rest and comparing with the rest of the rest and
so on was perfectly modeled by recursing over the list.

Finally, I was able to skip even more comparisons by returning early
once a match was found.